Rename contribution widget and fix tests
